Purpose:
hzip update

Description:

implemented copying of vdata and vdata fields to the new file

enabled the writing of more objects to the input file


Solution:

Platforms tested:

linux
SGI (modi4)
Sun 5.7 (arabica)



    [machines you have tested the changed version.  This is absolute
    important.  Test it out on at least two or three different platforms
    such as Big-endian-32bit (SUN/IRIX), little-endian-32(LINUX) and
    64-bit (IRIX64/UNICOS/DEC-ALPHA) would be good.]

+++ mfhdf/hzip/parse.c
@@ -84,7 +84,7 @@ obj_list_t* parse_comp(char *str, int *n_objs, comp_info_t *comp)
   }
  }
  /* nothing after : */
- if (end_obj+1==len)
+ if (end_obj+1==(int)len)
  {
   if (obj_list) free(obj_list);
   printf("%s\nError: Invalid compression type\n",str);
@@ -221,7 +221,7 @@ obj_list_t* parse_chunk(char *str, int *n_objs, int32 *chunk_lengths, int *chunk
  }
 
  /* nothing after : */
- if (end_obj+1==len)
+ if (end_obj+1==(int)len)
  {
   if (obj_list) free(obj_list);
   printf("%s\nError: Invalid chunking\n",str);
